[FAB-10865] ChaincodeStub does not work w/ TypeScript
The newly added TypeScript definitions for the fabric-shim package
reference a ChaincodeStub class, but this is exported as Stub. This
means that the TypeScript compiles successfully, but it fails at
runtime. Need to export ChaincodeStub as both Stub (for compatibility
reasons) and ChaincodeStub).

Also, adding in the RESPONSE_CODE constants.

Change-Id: I14c19c8e6a5f8869be382e47ed67521bfa30ac31
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>