[FAB-13250] Enable auto publish of new modules/tags
The current CI scripts for publishing to npm fail if
the module does not exist in npm (e.g. newly added
module) or if the tag for the module does not exist
in npm (e.g. newly added tag for new version). This
change updates the scripts to handle both of these
cases and default to "x.y.z-snapshot.1" if there is
no module or tag in npm.

Change-Id: I2d48d7484c60906bf73fa1b4086ab00ba75c0dec
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>