[FAB-16473] Copy proto files into source control
This is a temporary fix to get the builds working again.

After we're back to a working state, I plan on updating
the code to use protobufjs 6, so that we can just pull
the latest protos in again. The patch-package hack Gari
put into the SDK won't work here, as we'd need to patch
the code on the users system.

The ideal solution is to move the fabric-protos module
out of fabric-sdk-node and integrate it into the new
fabric-protos repository, but apparently any additional
updates to that CI is blocked for now:
"we are not building more into github/azure until after 2.0"

I don't really want to add a chaincode -> SDK dependency
as that is bound to cause us issues come release time.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I7c868a9f1e31decb93e5fd89819638b84392b965